+++ project/src/components/common/card.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Offer } from '../../types/types';
-import { FAVORITES } from '../../utils/const';
+import { FAVORITES } from '../../const';
 import FavoriteButton from './favorite-button';
 import Premium from './premium';
 import Price from './price';
