+++ project/src/components/main/sort.tsx
@@ -2,7 +2,7 @@ import { Dispatch, useState } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { changeSortName } from '../../store/action';
 import { State, Actions } from '../../types/reducer';
-import { sortTypes } from '../../utils/const';
+import { sortTypes } from '../../const';
 
 const mapStateToProps = ({ sortName }: State) => ({
   sortName,
