+++ project/src/components/properties/review-form/comments.tsx
@@ -1,5 +1,5 @@
 import Image from '../../common/img-avatar';
-import { dateFormatter } from '../../../utils/const';
+import { dateFormatter } from '../../../const';
 import { Review } from '../../../types/types';
 import { transformRating } from '../../../utils/utils';
 import UserName from '../../common/user-name';
