+++ project/src/const.ts
@@ -1,4 +1,4 @@
-import { Offer } from '../types/types';
+import { Offer } from './types/types';
 
 const months = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
