--- project/src/hooks/Readme.md
@@ -1,3 +1,0 @@
-# hooks
-
-Кастомные хуки. Каждый хук в отдельном модуле.
